+++ Sources/MapboxDirections/Extensions/Codable.swift
@@ -1,6 +1,7 @@
 import Foundation
 import Polyline
-import struct Turf.LineString
+import Turf
+import CoreLocation
 
 extension LineString {
     /**
@@ -28,12 +29,18 @@ enum PolyLineString {
 }
 
 extension PolyLineString: Codable {
+    private enum LineStringCodingKeys: String, CodingKey {
+        case coordinates
+    }
+    
     init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
         let options = decoder.userInfo[.options] as? DirectionsOptions
         switch options?.shapeFormat ?? .default {
         case .geoJSON:
-            self = .lineString(try container.decode(LineString.self))
+            let lineStringContainer = try decoder.container(keyedBy: LineStringCodingKeys.self)
+            let coordinates = try lineStringContainer.decode([CLLocationCoordinate2DCodable].self, forKey: .coordinates).map { $0.decodedCoordinates }
+            self = .lineString(LineString(coordinates))
         case .polyline, .polyline6:
             let precision = options?.shapeFormat == .polyline6 ? 1e6 : 1e5
             let encodedPolyline = try container.decode(String.self)
@@ -45,9 +52,42 @@ extension PolyLineString: Codable {
         var container = encoder.singleValueContainer()
         switch self {
         case let .lineString(lineString):
-            try container.encode(lineString)
+            var lineStringContainer = encoder.container(keyedBy: LineStringCodingKeys.self)
+            try lineStringContainer.encode(lineString.coordinates.map { CLLocationCoordinate2DCodable($0) }, forKey: .coordinates)
         case let .polyline(encodedPolyline, precision: _):
             try container.encode(encodedPolyline)
         }
+    }
+}
+
+struct CLLocationCoordinate2DCodable: Codable {
+    var latitude: CLLocationDegrees
+    var longitude: CLLocationDegrees
+    var decodedCoordinates: CLLocationCoordinate2D {
+        return CLLocationCoordinate2D(latitude: latitude,
+                                      longitude: longitude)
+    }
+    
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.unkeyedContainer()
+        try container.encode(longitude)
+        try container.encode(latitude)
+    }
+    
+    init(from decoder: Decoder) throws {
+        var container = try decoder.unkeyedContainer()
+        longitude = try container.decode(CLLocationDegrees.self)
+        latitude = try container.decode(CLLocationDegrees.self)
+    }
+    
+    init(_ coordinate: CLLocationCoordinate2D) {
+        latitude = coordinate.latitude
+        longitude = coordinate.longitude
+    }
+}
+
+extension CLLocationCoordinate2D {
+    var codableCoordinates: CLLocationCoordinate2DCodable {
+        return CLLocationCoordinate2DCodable(self)
     }
 }