+++ Sources/MapboxDirections/MapMatching/Match.swift
@@ -1,11 +1,44 @@
 import Foundation
 import CoreLocation
 import Polyline
-import struct Turf.LineString
+import Turf
 
-extension CodingUserInfoKey {
-    static let tracepoints = CodingUserInfoKey(rawValue: "com.mapbox.directions.coding.tracepoints")!
+/**
+ A `Weight` enum represents the weight given to a specific `Match` by the Directions API. The default metric is a compound index called "routability", which is duration-based with additional penalties for less desirable maneuvers.
+ */
+public enum Weight: Equatable {
+    
+    case routability(value: Float)
+    case other(value: Float, metric: String)
+    
+    public init(value: Float, metric: String) {
+        switch metric {
+        case "routability":
+            self = .routability(value: value)
+        default:
+            self = .other(value: value, metric: metric)
+        }
+    }
+    
+    var metric: String {
+        switch self {
+        case .routability(value: _):
+            return "routability"
+        case let .other(value: _, metric: value):
+            return value
+        }
+    }
+    
+    var value: Float {
+        switch self {
+        case let .routability(value: weight):
+            return weight
+        case let .other(value: weight, metric: _):
+            return weight
+        }
+    }
 }
+
 /**
  A `Match` object defines a single route that was created from a series of points that were matched against a road network.
  
@@ -14,8 +47,8 @@ extension CodingUserInfoKey {
 open class Match: DirectionsResult {
     private enum CodingKeys: String, CodingKey {
         case confidence
-        case tracepoints
-        case matchOptions
+        case weight
+        case weightName = "weight_name"
     }
     
     /**
@@ -28,14 +61,12 @@ open class Match: DirectionsResult {
      - parameter distance: The matched path’s cumulative distance, measured in meters.
      - parameter expectedTravelTime: The route’s expected travel time, measured in seconds.
      - parameter confidence: A number between 0 and 1 that indicates the Map Matching API’s confidence that the match is accurate. A higher confidence means the match is more likely to be accurate.
-     - parameter tracepoints: Tracepoints on the road network that match the tracepoints in `options`.
-     - parameter options: The criteria to match.
+     - parameter weight: A `Weight` enum, which represents the weight given to a specific `Match`.
      */
-    public init(legs: [RouteLeg], shape: LineString?, distance: CLLocationDistance, expectedTravelTime: TimeInterval, confidence: Float, tracepoints: [Tracepoint?], options: MatchOptions) {
-        matchOptions = options
+    public init(legs: [RouteLeg], shape: LineString?, distance: CLLocationDistance, expectedTravelTime: TimeInterval, confidence: Float, weight: Weight) {
         self.confidence = confidence
-        self.tracepoints = tracepoints
-        super.init(legs: legs, shape: shape, distance: distance, expectedTravelTime: expectedTravelTime, options: options)
+        self.weight = weight
+        super.init(legs: legs, shape: shape, distance: distance, expectedTravelTime: expectedTravelTime)
     }
     
     /**
@@ -47,44 +78,33 @@ open class Match: DirectionsResult {
     public required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         confidence = try container.decode(Float.self, forKey: .confidence)
-        tracepoints = try container.decodeIfPresent([Tracepoint?].self, forKey: .tracepoints) ?? []
-        if let matchOptions = try container.decodeIfPresent(MatchOptions.self, forKey: .matchOptions)
-            ?? decoder.userInfo[.options] as? MatchOptions {
-            self.matchOptions = matchOptions
-        } else {
-            throw DirectionsCodingError.missingOptions
-        }
+        let weightValue = try container.decode(Float.self, forKey: .weight)
+        let weightMetric = try container.decode(String.self, forKey: .weightName)
+        
+        weight = Weight(value: weightValue, metric: weightMetric)
+        
         try super.init(from: decoder)
     }
     
     public override func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(confidence, forKey: .confidence)
-        try container.encode(tracepoints, forKey: .tracepoints)
-        try container.encode(matchOptions, forKey: .matchOptions)
+        try container.encode(weight.value, forKey: .weight)
+        try container.encode(weight.metric, forKey: .weightName)
+        
         try super.encode(to: encoder)
     }
+    
+    /**
+     A `Weight` enum, which represents the weight given to a specific `Match`.
+     */
+    open var weight: Weight
     
     /**
      A number between 0 and 1 that indicates the Map Matching API’s confidence that the match is accurate. A higher confidence means the match is more likely to be accurate.
      */
     open var confidence: Float
     
-    /**
-     Tracepoints on the road network that match the tracepoints in the match options.
-     
-     Any outlier tracepoint is omitted from the match. This array represents an outlier tracepoint if the element is `nil`.
-     */
-    open var tracepoints: [Tracepoint?]
-    
-    public override var directionsOptions: DirectionsOptions {
-        return matchOptions
-    }
-    
-    /**
-     `MatchOptions` used to create the match request.
-     */
-    public let matchOptions: MatchOptions
 }
 
 extension Match: Equatable {
@@ -93,8 +113,9 @@ extension Match: Equatable {
             lhs.distance == rhs.distance &&
             lhs.expectedTravelTime == rhs.expectedTravelTime &&
             lhs.speechLocale == rhs.speechLocale &&
+            lhs.responseContainsSpeechLocale == rhs.responseContainsSpeechLocale &&
             lhs.confidence == rhs.confidence &&
-            lhs.tracepoints == rhs.tracepoints &&
+            lhs.weight == rhs.weight &&
             lhs.legs == rhs.legs &&
             lhs.shape == rhs.shape
     }